requirements.txt
a2a-sdk==0.2.8
annotated-types==0.7.0
anyio==4.9.0
authlib==1.6.0
bleach==6.2.0
bokeh==3.7.3
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
dask==2025.5.1
distributed==2025.5.1
distro==1.9.0
docstring-parser==0.16
faker==37.4.0
fastapi==0.115.12
filetype==1.2.0
fonttools==4.58.4
fsspec==2025.5.1
ftfy==6.3.1
google-adk==1.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.97.0
google-cloud-appengine-logging==1.6.2
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.20.0
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.21
grpc-google-iam-v1==0.14.2
grpcio==1.73.0
grpcio-reflection==1.71.0
grpcio-status==1.71.0
grpcio-tools==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib-metadata==8.7.0
iniconfig==2.1.0
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
kaggle==1.7.4.5
-e file:///Users/prasanjitdutta/Desktop/projects2/datateam/kaggle-mcp
kiwisolver==1.4.8
langchain==0.3.25
langchain-core==0.3.65
langchain-google-genai==2.0.10
langchain-openai==0.3.23
langchain-text-splitters==0.3.8
langgraph==0.4.8
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.2
langgraph-sdk==0.1.70
langsmith==0.3.45
locket==1.0.0
lz4==4.4.4
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.3
mcp==1.9.4
mcp-server==0.1.4
mdurl==0.1.2
missingno==0.5.2
msgpack==1.1.1
multipledispatch==1.0.0
narwhals==1.42.1
natsort==8.4.0
numpy==2.3.1
openai==1.86.0
opentelemetry-api==1.34.1
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.10.18
ormsgpack==1.10.0
packaging==25.0
panda==0.3.1
pandas==2.3.0
pandas-flavor==0.7.0
partd==1.4.2
phonenumbers==9.0.7
pillow==11.2.1
pluggy==1.6.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.9.1
pygments==2.19.1
pyjanitor==0.31.0
pyparsing==3.2.3
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlalchemy==2.0.41
sse-starlette==2.3.6
starlette==0.46.2
tblib==3.1.0
tenacity==9.1.2
text-unidecode==1.3
threadpoolctl==3.6.0
tiktoken==0.9.0
toolz==1.0.0
tornado==6.5.1
tqdm==4.67.1
typer==0.16.0
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
unidecode==1.4.0
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
xarray==2025.6.1
xxhash==3.5.0
xyzservices==2025.4.0
zict==3.0.0
zipp==3.23.0
zstandard==0.23.0
